Add `ask-open-desktop` param to ZDS link
Since our implementation of temporary workspaces in browser Zoo Design
Studio as of https://github.com/KittyCAD/modeling-app/pull/7393, we need
to add this query parameter to links like this one so that it can pick
up on the temporary workspace logic. We can upstream a fix to let the
temporary workspace logic run without it from the ZDS side separately,
but this is a no-harm fix in the meantime.