Fix /api/feedback 500 error
This endpoint actually works every time, but it also currently 500s every time. It's returning 500 because it's trying to load the response from our API as json, but we actually don't return anything in the body. So, api-deux succeeds, returns empty body, then this endpoint fails. But, no error is ever shown to the user, so it doesn't appear to be broken, it's just coincidental that it works fine.

I don't think we consume the message in the response anywhere anyway, so I just removed that entirely.